Create separate module for visualizing neat genomes
