+++ nevopy/neat/genomes.py
@@ -30,38 +30,20 @@ the network it encodes. In NEAT, the genome is the entity subject to evolution.
 
 import logging
 import os
-from numbers import Number
-from typing import Any, cast, Dict, List, Optional, Sequence, Tuple, Union
-from typing import TYPE_CHECKING
+from typing import Any, cast, Dict, List, Optional, Sequence, Tuple
 
-import matplotlib.pyplot as plt
 import numpy as np
 np.warnings.filterwarnings("ignore", category=np.VisibleDeprecationWarning) \
     # pylint: disable=wrong-import-position
-import networkx as nx
 from tensorflow import reshape
 
-import nevopy.activations as activations
-import nevopy.utils as utils
-from nevopy.base_genome import BaseGenome
-from nevopy.base_genome import InvalidInputError
-from nevopy.base_genome import IncompatibleGenomesError
-from nevopy.neat.genes import align_connections
-from nevopy.neat.genes import ConnectionGene
-from nevopy.neat.genes import NodeGene
-from nevopy.neat.config import NeatConfig
-from nevopy.neat.id_handler import IdHandler
-from nevopy.fixed_topology.genomes import FixedTopologyGenome
+import nevopy as ne
 
 _logger = logging.getLogger(__name__)
 os.environ["PYGAME_HIDE_SUPPORT_PROMPT"] = "hide"
 
-# Necessary for forward-reference type-checking.
-if TYPE_CHECKING:
-    import pygame
 
-
-class NeatGenome(BaseGenome):
+class NeatGenome(ne.base_genome.BaseGenome):
     """ Linear representation of a neural network's connectivity.
 
     In the context of NEAT, a genome is a collection of genes that encode a
@@ -110,7 +92,7 @@ class NeatGenome(BaseGenome):
     def __init__(self,
                  num_inputs: int,
                  num_outputs: int,
-                 config: NeatConfig,
+                 config: "ne.neat.config.NeatConfig",
                  initial_connections: bool = True) -> None:
         super().__init__()
         self._config = config
@@ -120,34 +102,35 @@ class NeatGenome(BaseGenome):
         self.fitness = 0.0
         self.adj_fitness = 0.0
 
-        self._input_nodes = []
-        self.hidden_nodes = []  # type: List[NodeGene]
-        self._output_nodes = []
-        self._bias_node = None
+        self.input_nodes = []   # type: List["ne.neat.NodeGene"]
+        self.hidden_nodes = []  # type: List["ne.neat.NodeGene"]
+        self.output_nodes = []  # type: List["ne.neat.NodeGene"]
+        self.bias_node = None   # type: Optional["ne.neat.NodeGene"]
 
-        self.connections = []  # type: List[ConnectionGene]
+        self.connections = []  # type: List["ne.neat.ConnectionGene"]
         self._existing_connections_dict = {} \
-            # type: Dict[int, Dict[int, ConnectionGene]]
+            # type: Dict[int, Dict[int, "ne.neat.ConnectionGene"]]
         self._output_activation = self.config.out_nodes_activation
         self._hidden_activation = self.config.hidden_nodes_activation
 
         # init input nodes
         node_counter = 0
         for _ in range(num_inputs):
-            self._input_nodes.append(
-                NodeGene(node_id=node_counter,
-                         node_type=NodeGene.Type.INPUT,
-                         activation_func=activations.linear,
-                         initial_activation=self.config.initial_node_activation)
+            self.input_nodes.append(
+                ne.neat.NodeGene(
+                    node_id=node_counter,
+                    node_type=ne.neat.NodeGene.Type.INPUT,
+                    activation_func=ne.activations.linear,
+                    initial_activation=self.config.initial_node_activation)
             )
             node_counter += 1
 
         # init bias node
         if self.config.bias_value is not None:
-            self._bias_node = NodeGene(
+            self.bias_node = ne.neat.NodeGene(
                 node_id=node_counter,
-                node_type=NodeGene.Type.BIAS,
-                activation_func=activations.linear,
+                node_type=ne.neat.NodeGene.Type.BIAS,
+                activation_func=ne.activations.linear,
                 initial_activation=self.config.bias_value,
             )
             node_counter += 1
@@ -155,30 +138,30 @@ class NeatGenome(BaseGenome):
         # init output nodes
         connection_counter = 0
         for _ in range(num_outputs):
-            out_node = NodeGene(
+            out_node = ne.neat.NodeGene(
                 node_id=node_counter,
-                node_type=NodeGene.Type.OUTPUT,
+                node_type=ne.neat.NodeGene.Type.OUTPUT,
                 activation_func=self._output_activation,
                 initial_activation=self.config.initial_node_activation,
             )
-            self._output_nodes.append(out_node)
+            self.output_nodes.append(out_node)
             node_counter += 1
 
             # connecting all input nodes to all output nodes
             if initial_connections:
-                for in_node in self._input_nodes:
+                for in_node in self.input_nodes:
                     connection_counter += 1
                     self.add_connection(connection_counter, in_node, out_node)
 
     @property
     def input_shape(self) -> int:
         """ Number of input nodes in the genome. """
-        return len(self._input_nodes)
+        return len(self.input_nodes)
 
     @property
     def output_shape(self) -> int:
         """ Number of output nodes in the genome. """
-        return len(self._output_nodes)
+        return len(self.output_nodes)
 
     @property
     def config(self) -> Any:
@@ -232,7 +215,7 @@ class NeatGenome(BaseGenome):
         Returns:
             The distance between the genomes.
         """
-        genes = align_connections(self.connections, other.connections)
+        genes = ne.neat.align_connections(self.connections, other.connections)
         excess = disjoint = num_matches = 0
         weight_diff = 0.0
 
@@ -280,8 +263,8 @@ class NeatGenome(BaseGenome):
 
     def add_connection(self,
                        cid: int,
-                       src_node: NodeGene,
-                       dest_node: NodeGene,
+                       src_node: "ne.neat.genes.NodeGene",
+                       dest_node: "ne.neat.genes.NodeGene",
                        enabled: bool = True,
                        weight: Optional[float] = None) -> None:
         """ Adds a new connection gene to the genome.
@@ -308,8 +291,8 @@ class NeatGenome(BaseGenome):
             raise ConnectionExistsError(
                 f"Attempt to create an already existing connection "
                 f"({src_node.id}->{dest_node.id}).")
-        if (dest_node.type == NodeGene.Type.BIAS
-                or dest_node.type == NodeGene.Type.INPUT):
+        if (dest_node.type == ne.neat.NodeGene.Type.BIAS
+                or dest_node.type == ne.neat.NodeGene.Type.INPUT):
             raise ConnectionToBiasNodeError(
                 f"Attempt to create a connection pointing to a bias or input "
                 f"node ({src_node.id}->{dest_node.id}). Nodes of this type "
@@ -317,10 +300,10 @@ class NeatGenome(BaseGenome):
 
         weight = (np.random.uniform(*self.config.new_weight_interval)
                   if weight is None else weight)
-        connection = ConnectionGene(cid=cid,
-                                    from_node=src_node,
-                                    to_node=dest_node,
-                                    weight=weight)
+        connection = ne.neat.ConnectionGene(cid=cid,
+                                            from_node=src_node,
+                                            to_node=dest_node,
+                                            weight=weight)
 
         connection.enabled = enabled
         self.connections.append(connection)
@@ -332,8 +315,8 @@ class NeatGenome(BaseGenome):
         self._existing_connections_dict[src_node.id][dest_node.id] = connection
 
     def add_random_connection(self,
-                              id_handler: IdHandler,
-    ) -> Optional[Tuple[NodeGene, NodeGene]]:
+                              id_handler: "ne.neat.id_handler.IdHandler",
+    ) -> Optional[Tuple["ne.neat.genes.NodeGene", "ne.neat.genes.NodeGene"]]:
         """  Adds a new connection between two random nodes in the genome.
 
         This is an implementation of the `add connection mutation`, described in
@@ -353,8 +336,8 @@ class NeatGenome(BaseGenome):
         np.random.shuffle(all_src_nodes)
 
         all_dest_nodes = [n for n in all_src_nodes
-                          if (n.type != NodeGene.Type.BIAS
-                              and n.type != NodeGene.Type.INPUT)]
+                          if (n.type != ne.neat.NodeGene.Type.BIAS
+                              and n.type != ne.neat.NodeGene.Type.INPUT)]
         np.random.shuffle(all_dest_nodes)
 
         for src_node in all_src_nodes:
@@ -375,7 +358,8 @@ class NeatGenome(BaseGenome):
             connection.enabled = True
 
     def add_random_hidden_node(self,
-                               id_handler: IdHandler) -> Optional[NodeGene]:
+                               id_handler: "ne.neat.id_handler.IdHandler",
+    ) -> Optional["ne.neat.genes.NodeGene"]:
         """ Adds a new hidden node to the genome in a random position.
 
         This method implements the `add node mutation` procedure described in
@@ -416,9 +400,9 @@ class NeatGenome(BaseGenome):
                 continue
 
             original_connection.enabled = False
-            new_node = NodeGene(
+            new_node = ne.neat.NodeGene(
                 node_id=hid,
-                node_type=NodeGene.Type.HIDDEN,
+                node_type=ne.neat.NodeGene.Type.HIDDEN,
                 activation_func=self._hidden_activation,
                 initial_activation=self.config.initial_node_activation
             )
@@ -445,7 +429,7 @@ class NeatGenome(BaseGenome):
         or to remain unchanged.
         """
         for connection in self.connections:
-            if utils.chance(self.config.weight_reset_chance):
+            if ne.utils.chance(self.config.weight_reset_chance):
                 # perturbating the connection
                 connection.weight = np.random.uniform(
                     *self.config.new_weight_interval)
@@ -465,8 +449,8 @@ class NeatGenome(BaseGenome):
             A copy of the genome without any of its connections (including the
             ones between input and output nodes) and hidden nodes.
         """
-        return NeatGenome(num_inputs=len(self._input_nodes),
-                          num_outputs=len(self._output_nodes),
+        return NeatGenome(num_inputs=len(self.input_nodes),
+                          num_outputs=len(self.output_nodes),
                           config=self.config,
                           initial_connections=False)
 
@@ -521,7 +505,7 @@ class NeatGenome(BaseGenome):
         """
         return self.__copy_aux(random_weights=False)
 
-    def process_node(self, n: NodeGene) -> float:
+    def process_node(self, n: "ne.neat.genes.NodeGene") -> float:
         """ Recursively processes the activation of the given node.
 
         Unless it's a bias or input node (that have a fixed output), a node must
@@ -547,8 +531,8 @@ class NeatGenome(BaseGenome):
             The activation value (output) of the node.
         """
         # checking if the node needs to be activated
-        if (n.type != NodeGene.Type.INPUT
-                and n.type != NodeGene.Type.BIAS
+        if (n.type != ne.neat.NodeGene.Type.INPUT
+                and n.type != ne.neat.NodeGene.Type.BIAS
                 and not self._activated_nodes[n.id]):
             # activating the node
             # the current node (n) is immediately marked as activated; this is
@@ -593,39 +577,39 @@ class NeatGenome(BaseGenome):
             InvalidInputError: If the number of elements in `X` doesn't match
                 the number of input nodes in the network.
         """
-        if len(x) != len(self._input_nodes):
-            raise InvalidInputError(
+        if len(x) != len(self.input_nodes):
+            raise ne.InvalidInputError(
                 "The input size must match the number of input nodes in the "
-                f"network! Expected input of length {len(self._input_nodes)} "
+                f"network! Expected input of length {len(self.input_nodes)} "
                 f"but got {len(x)}."
             )
 
         # preparing input nodes
-        for in_node, value in zip(self._input_nodes, x):
+        for in_node, value in zip(self.input_nodes, x):
             in_node.activate(value)
 
         # resetting activated nodes dict
         self._activated_nodes = {
             n.id: False
-            for n in self._output_nodes + self.hidden_nodes
+            for n in self.output_nodes + self.hidden_nodes
         }
 
         # processing nodes in a top-down manner (starts from the output nodes)
         # nodes not connected to at least one output node are not processed
-        h = np.zeros(len(self._output_nodes))
-        for i, out_node in enumerate(self._output_nodes):
+        h = np.zeros(len(self.output_nodes))
+        for i, out_node in enumerate(self.output_nodes):
             h[i] = self.process_node(out_node)
 
         return h
 
-    def nodes(self) -> List[NodeGene]:
+    def nodes(self) -> List["ne.neat.genes.NodeGene"]:
         """
         Returns all the genome's node genes. Order: inputs, bias, outputs and
         hidden.
         """
-        return (self._input_nodes +
-                ([self._bias_node] if self._bias_node is not None else []) +
-                self._output_nodes +
+        return (self.input_nodes +
+                ([self.bias_node] if self.bias_node is not None else []) +
+                self.output_nodes +
                 self.hidden_nodes)
 
     def valid_out_nodes(self) -> bool:
@@ -641,7 +625,7 @@ class NeatGenome(BaseGenome):
             incoming connection and `False` otherwise. Self-connecting
             connections are not considered.
         """
-        for out_node in self._output_nodes:
+        for out_node in self.output_nodes:
             valid = False
             for in_con in out_node.in_connections:
                 if in_con.enabled and not in_con.self_connecting():
@@ -662,7 +646,7 @@ class NeatGenome(BaseGenome):
             `True` if all the genome's input nodes are valid and `False`
             otherwise.
         """
-        for in_node in self._input_nodes:
+        for in_node in self.input_nodes:
             valid = False
             for out_con in in_node.out_connections:
                 if out_con.enabled:
@@ -701,14 +685,14 @@ class NeatGenome(BaseGenome):
                 ``other`` is incompatible with the current genome (`self`).
         """
         if not issubclass(type(other), NeatGenome):
-            raise IncompatibleGenomesError(
+            raise ne.IncompatibleGenomesError(
                 "Instances of `NeatGenome` are currently only compatible for "
                 "sexual reproduction with instances of `NeatGenome or one of "
                 "its subclasses!"
             )
 
         # aligning matching genes
-        genes = align_connections(self.connections, other.connections)
+        genes = ne.neat.align_connections(self.connections, other.connections)
 
         # new genome
         new_gen = self.simple_copy()
@@ -740,13 +724,13 @@ class NeatGenome(BaseGenome):
                 enabled = True
                 if ((c1 is not None and not c1.enabled)
                         or (c2 is not None and not c2.enabled)):
-                    enabled = not utils.chance(
+                    enabled = not ne.utils.chance(
                         self.config.disable_inherited_connection_chance)
                 chosen_connections.append((c, enabled))
 
                 # adding the hidden nodes of the connection (if needed)
                 for node in (c.from_node, c.to_node):
-                    if (node.type == NodeGene.Type.HIDDEN
+                    if (node.type == ne.neat.NodeGene.Type.HIDDEN
                             and node.id not in copied_nodes):
                         new_node = node.simple_copy()
                         new_gen.hidden_nodes.append(new_node)
@@ -785,659 +769,19 @@ class NeatGenome(BaseGenome):
                    f"[{c.from_node.id}->{c.to_node.id}] {c.weight}\n"
         return txt
 
-    def columns_graph_layout(self,
-                             width: float,
-                             height: float,
-                             node_size: float,
-                             horizontal_pad_pc: Tuple[float,
-                                                      float] = (.03, .03),
-                             vertical_pad_pc: Tuple[float,
-                                                    float] = (.03, .03),
-                             ideal_h_nodes_per_col: int = 4,
-                             consider_bias_node: bool = True,
-    ) -> Dict[int, Tuple[float, float]]:
-        """ Positions the network's nodes in columns.
-
-        The input nodes are placed in the left-most column and the output nodes
-        are placed in the right-most columns. The hidden nodes are placed in
-        columns located between those two columns. For big networks, try using a
-        smaller node size for better quality.
-
-        Args:
-            width (float): Width of the figure / surface.
-            height (float): Height of the figure / surface.
-            node_size (float): Size of the drawn nodes.
-            horizontal_pad_pc (Tuple[float, float]): Tuple containing the size
-                of the padding on the left and on the right of the surface.
-                Unit: the width of the surface.
-            vertical_pad_pc (Tuple[float, float]): Tuple containing the size
-                of the padding below and above the surface. Unit: the height of
-                the surface.
-            ideal_h_nodes_per_col (int): Preferred number of hidden nodes per
-                column (the algorithm will try to draw columns with this amount
-                of hidden nodes when possible).
-            consider_bias_node (bool): Whether the bias node should be
-                considered or not when calculating the positions.
-
-        Returns:
-            Dictionary mapping the ID of each node to a tuple containing its
-            position in the figure.
+    def visualize(self, **kwargs) -> None:
+        """ Simple wrapper for the
+        :func:`nevopy.neat.visualization.visualize_genome` function. Please
+        refer to its documentation for more information.
         """
-        pos = {}
+        ne.neat.visualize_genome(genome=self, **kwargs)
 
-        # padding
-        origin_x = width * horizontal_pad_pc[0] + node_size / 2
-        origin_y = height * vertical_pad_pc[0] + node_size / 2
-
-        width = width - origin_x - horizontal_pad_pc[1] * width - node_size / 2
-        height = height - origin_y - vertical_pad_pc[1] * height - node_size / 2
-
-        # procedure for inserting nodes into columns
-        def insert_nodes_col(x, nodes):
-            """ Inserts the nodes in a column (specified by x). """
-            if len(nodes) == 1:
-                pos[nodes[0].id] = (x, origin_y + height/2)
-                return
-
-            next_y = origin_y
-            space_y = height / len(nodes)
-            for n in nodes:
-                pos[n.id] = (x, next_y + space_y/2)
-                next_y += space_y
-
-        # input and bias nodes
-        insert_nodes_col(
-            x=origin_x,
-            nodes=self._input_nodes + ([] if (self._bias_node is None
-                                              or not consider_bias_node)
-                                       else [self._bias_node]),
-        )
-
-        # output nodes
-        insert_nodes_col(x=origin_x + width,
-                         nodes=self._output_nodes)
-
-        # hidden nodes
-        if self.hidden_nodes:
-            max_num_h_cols = int((width - 4 * node_size) / (2 * node_size))
-            h_nodes_per_col = ideal_h_nodes_per_col
-            num_cols = max(1, np.ceil(len(self.hidden_nodes) / h_nodes_per_col))
-            while num_cols > max_num_h_cols:
-                h_nodes_per_col += 1
-                num_cols = np.ceil(len(self.hidden_nodes) / h_nodes_per_col)
-
-            if num_cols == 1:
-                insert_nodes_col(x=width / 2, nodes=self.hidden_nodes)
-            else:
-                next_x = origin_x + 2 * node_size
-                space_x = (width - 4 * node_size) / num_cols
-                h_nodes = np.array(self.hidden_nodes)
-                for node_list in np.array_split(h_nodes, num_cols):
-                    insert_nodes_col(x=next_x + space_x / 2,
-                                     nodes=node_list)
-                    next_x += space_x
-
-        return pos
-
-    def _nodes_activation_status(self,
-                                 output_activation_threshold: float,
-                                 hidden_activation_threshold: float,
-                                 input_activation_threshold: Union[float,
-                                                                   List[float]],
-    ) -> Dict[int, bool]:
-        """ Check each of the network's nodes activation status.
-
-        Returns:
-            A :class:`dict` mapping each node's ID to a :class:`bool` (``True``
-            if the node is activated and ``False`` otherwise).
+    def visualize_activations(self, **kwargs) -> Any:
+        """ Simple wrapper for the
+        :func:`nevopy.neat.visualization.visualize_activations` function. Please
+        refer to its documentation for more information.
         """
-        status = {}  # type: Dict[int, bool]
-        max_out_node = max(self._output_nodes, key=lambda n: n.activation)
-
-        for node in self.nodes():
-            # Bias node:
-            if node.type == NodeGene.Type.BIAS:
-                status[node.id] = False
-                continue
-
-            # Output node:
-            if node.type == NodeGene.Type.OUTPUT:
-                if len(self._output_nodes) > 1:
-                    status[node.id] = node is max_out_node
-                    continue
-                threshold = output_activation_threshold
-            # Input node:
-            elif node.type == node.type.INPUT:
-                threshold = (
-                    input_activation_threshold
-                    if isinstance(input_activation_threshold, Number)
-                    else input_activation_threshold.pop(0)  # type: ignore
-                )
-            # Hidden node:
-            else:
-                threshold = hidden_activation_threshold
-
-            status[node.id] = node.activation > threshold
-
-        return status
-
-    # noinspection PyUnboundLocalVariable
-    def visualize_activations(self,
-                              # Sizes
-                              surface_size: Tuple[int, int] = (700, 450),
-                              node_radius: float = 14,
-                              # Nodes colors
-                              node_deactivated_color: Union[
-                                  str, Tuple[int, int, int]] = (190, 190, 190),
-                              node_activated_color: Union[
-                                  str, Tuple[int, int, int]] = (2, 68, 144),
-                              bias_node_color: Union[
-                                  str, Tuple[int, int, int]] = "yellow",
-                              node_border_color: Union[
-                                  str, Tuple[int, int, int]] = "black",
-                              # Edges colors
-                              edge_activated_color: Union[
-                                  str, Tuple[int, int, int]] = (0, 120, 233),
-                              edge_deactivated_color: Union[
-                                  str, Tuple[int, int, int]] = (100, 100, 100),
-                              # Edges width
-                              activated_edge_width: int = 2,
-                              deactivated_edge_width: int = 1,
-                              # Padding
-                              horizontal_pad_pc: Tuple[float,
-                                                       float] = (.015, .015),
-                              vertical_pad_pc: Tuple[float,
-                                                     float] = (.015, .015),
-                              # Activation threshold
-                              hidden_activation_threshold: float = 0.5,
-                              input_activation_threshold: Union[
-                                  float, List[float]] = 0.5,
-                              output_activation_threshold: float = 0.5,
-                              # Labels
-                              input_labels: Optional[List[str]] = None,
-                              output_labels: Optional[List[str]] = None,
-                              labels_color: Union[
-                                    str, Tuple[int, int, int]] = "white",
-                              labels_config: Dict[str, Any] = None,
-                              # Light
-                              show_activation_light: bool = True,
-                              activation_light_color: Optional[Union[
-                                  str, Tuple[int, int, int]]] = (104, 179, 235),
-                              activation_light_radius_pc: float = 2,
-                              # Others
-                              ideal_h_nodes_per_col: int = 4,
-                              background_color: Union[
-                                  str, Tuple[int, int, int]] = "black",
-                              node_border_thickness: Optional[float] = 2,
-                              draw_bias_node: bool = False,
-                              return_rgb_array: bool = False,
-    ) -> Union["pygame.surface.Surface", np.ndarray]:
-        """ Draws the network taking into consideration each node's activation.
-
-        Activated nodes and edges are drawn with different colors. This method
-        requires :mod:`pygame`.
-
-        Args:
-            surface_size (Tuple[int, int]): Width and height of the pygame
-                surface to be drawn.
-            node_radius (float): Radius (size) of the nodes.
-            node_deactivated_color (Union[str, Tuple[int, int, int]]): Color of
-                deactivated nodes.
-            node_activated_color (Union[str, Tuple[int, int, int]]): Color of
-                activated nodes.
-            bias_node_color (Union[str, Tuple[int, int, int]]): Color of the
-                bias node.
-            node_border_color (Union[str, Tuple[int, int, int]]): Color of the
-                nodes' borders.
-            edge_activated_color (Union[str, Tuple[int, int, int]]): Color of
-                activated edges.
-            edge_deactivated_color (Union[str, Tuple[int, int, int]]): Color of
-                deactivated edges.
-            horizontal_pad_pc (Tuple[float, float]): Tuple containing the size
-                of the padding on the left and on the right of the surface.
-                Unit: the width of the surface.
-            vertical_pad_pc (Tuple[float, float]): Tuple containing the size
-                of the padding below and above the surface. Unit: the height of
-                the surface.
-            hidden_activation_threshold (float): Activation threshold for hidden
-                nodes. If the activation value of a hidden node is greater than
-                this threshold, the node is considered to be activated.
-            input_activation_threshold (Union[float, List[float]]): Activation
-                threshold(s) for input nodes. If the argument passed to this
-                parameter is a `float`, the same threshold will be considered
-                for all the input nodes. If it's a list, each input node will
-                have its own activation threshold. If the activation value of an
-                input node is greater than the threshold, the node is considered
-                to be activated.
-            output_activation_threshold (float): Activation threshold for output
-                nodes. This value is only used when the network has only one
-                output node. When there are many output nodes, the activated
-                node is always the one with the greatest activation value.
-            input_labels (Optional[List[str]]): Labels for the input nodes.
-            output_labels (Optional[List[str]]): Labels for the output nodes.
-            labels_color (Union[str, Tuple[int, int, int]]): Color of the
-                labels.
-            labels_config (Dict[str, Any]): Keyword arguments to be passed to
-                the :meth:`pygame.SysFont` constructor.
-            show_activation_light (bool): Whether or not to show a "light ring"
-                around activated nodes.
-            activation_light_color (Optional[Union[str, Tuple[int, int, int]]]):
-                The color of the light ring to be shown around activated nodes.
-            activation_light_radius_pc (float): Radius of the light ring to be
-                drawn around activated nodes. Unit: the node's radius.
-            ideal_h_nodes_per_col (int): Preferred number of hidden nodes per
-                column (the algorithm will try to draw columns with this amount
-                of hidden nodes whenever possible).
-            background_color (Union[str, Tuple[int, int, int]]): The background
-                color of the surface.
-            node_border_thickness (Optional[float]): Thickness of the nodes'
-                borders. If ``None`` or `0`, no border will be drawn.
-            activated_edge_width (int): The width/thickness of activated edges.
-            deactivated_edge_width (int): The width/thickness of deactivated
-                edges.
-            draw_bias_node (bool): Whether to draw the network's bias node or
-                not.
-            return_rgb_array (bool): If ``True``, returns a numpy array with the
-                generated image instead of a pygame surface.
-
-        Returns:
-            If ``return_rgb_array`` is ``False``, an instance of
-            :class:`pygame.Surface` with the drawings is returned. You can
-            display it using :mod:`pygame`:
-
-                .. code:: python
-
-                    screen_size = 700, 450
-                    display = pygame.display.set_mode(screen_size)
-                    # ...
-                    s = genome.visualize_activations(surface_size=screen_size)
-                    display.blit(s, [0, 0])
-                    pygame.display.update()
-
-            If ``return_rgb_array`` is ``True``, a numpy array with the
-            generated image is returned instead.
-
-        Raises:
-            ModuleNotFoundError: If :mod:`pygame` is not found.
-        """
-        # Importing pygame:
-        try:
-            import pygame  # pylint: disable=import-outside-toplevel
-            if not pygame.font.get_init():
-                pygame.font.init()
-        except ModuleNotFoundError as e:
-            raise ModuleNotFoundError(
-                "Couldn't find 'pygame'! To use this method, make sure you've "
-                "it installed.\nYou can install 'pygame' using pip:\n"
-                "\t$ pip install pygame"
-            ) from e
-
-        # Creating surface:
-        surface = pygame.Surface(surface_size)
-        surface.fill(background_color)
-
-        # Font (for rendering the labels):
-        if input_labels is not None or output_labels is not None:
-            if labels_config is None:
-                labels_config = dict(name="monospace", size=15,
-                                     bold=True, italic=False)
-            font = pygame.font.SysFont(**labels_config)
-
-        # New variable for the horizontal padding:
-        # (required for placing the labels)
-        h_pad = list(horizontal_pad_pc)
-
-        # Rendering input labels:
-        if input_labels is not None:
-            rendered_in_labels = []  # type: List[pygame.surface.Surface]
-            for label_txt, node in zip(input_labels, self._input_nodes):
-                rendered_in_labels.append(
-                    font.render(f"{label_txt}: {node.activation:.2f}  -->   ",
-                                True, labels_color)  # type: ignore
-                )
-            max_width = max(rendered_in_labels,
-                            key=lambda s: s.get_size()[0]).get_size()[0]
-            h_pad[0] = (max_width + h_pad[0]*surface_size[0]) / surface_size[0]
-
-        # Rendering output labels:
-        if output_labels is not None:
-            rendered_out_labels = [font.render(" " * 4 + label_txt, True,
-                                               labels_color)  # type: ignore
-                                   for label_txt in output_labels]
-            max_width = max(rendered_out_labels,
-                            key=lambda s: s.get_size()[0]).get_size()[0]
-            h_pad[1] = (max_width + h_pad[1]*surface_size[1]) / surface_size[1]
-
-        # Checking the activation status of the nodes:
-        node_activated = self._nodes_activation_status(
-            output_activation_threshold=output_activation_threshold,
-            hidden_activation_threshold=hidden_activation_threshold,
-            input_activation_threshold=input_activation_threshold,
-        )
-
-        # Calculating the nodes' position:
-        nodes_pos = self.columns_graph_layout(
-            width=surface_size[0],
-            height=surface_size[1],
-            node_size=2 * node_radius,
-            horizontal_pad_pc=h_pad,  # type: ignore
-            vertical_pad_pc=vertical_pad_pc,
-            ideal_h_nodes_per_col=ideal_h_nodes_per_col,
-            consider_bias_node=draw_bias_node,
-        )
-
-        # Drawing input labels
-        if input_labels is not None:
-            for label, node in zip(rendered_in_labels, self._input_nodes):
-                x, y = nodes_pos[node.id]
-                w, h = label.get_size()
-                surface.blit(label, dest=(x - w, y - h / 2))
-
-        # Drawing input labels
-        if output_labels is not None:
-            for label, node in zip(rendered_out_labels, self._output_nodes):
-                x, y = nodes_pos[node.id]
-                h = label.get_size()[1]
-                surface.blit(label, dest=(x, y - h / 2))
-
-        # Drawing edges:
-        # (must be drawn before the nodes)
-        for c in self.connections:
-            if not c.enabled or (c.from_node.type == NodeGene.Type.BIAS
-                                 and not draw_bias_node):
-                continue
-
-            if node_activated[c.from_node.id] and c.weight > 0:
-                edge_color = edge_activated_color
-                edge_width = activated_edge_width
-            else:
-                edge_color = edge_deactivated_color
-                edge_width = deactivated_edge_width
-
-            pygame.draw.line(surface,
-                             color=edge_color,
-                             start_pos=nodes_pos[c.from_node.id],
-                             end_pos=nodes_pos[c.to_node.id],
-                             width=edge_width)
-
-        # Drawing nodes:
-        for node in self.nodes():
-            if node.type == NodeGene.Type.BIAS and not draw_bias_node:
-                continue
-
-            # Border:
-            if node_border_thickness is not None and node_border_thickness > 0:
-                pygame.draw.circle(surface,
-                                   color=node_border_color,
-                                   center=nodes_pos[node.id],
-                                   radius=node_radius + node_border_thickness)
-            # Choosing the node's color:
-            color = (bias_node_color if node.type == NodeGene.Type.BIAS
-                     else node_activated_color if node_activated[node.id]
-                     else node_deactivated_color)
-            # Drawing the node:
-            pygame.draw.circle(surface,
-                               color=color,
-                               center=nodes_pos[node.id],
-                               radius=node_radius)
-
-        # Drawing activation lights
-        if show_activation_light:
-            for node in self.nodes():
-                if node_activated[node.id]:
-                    # Making surface:
-                    light_radius = node_radius * activation_light_radius_pc
-                    light_surface = pygame.Surface(size=(2 * light_radius,
-                                                         2 * light_radius),
-                                                   flags=pygame.SRCALPHA)
-
-                    # Drawing the circle:
-                    pygame.draw.circle(light_surface,
-                                       color=(node_activated_color
-                                              if activation_light_color is None
-                                              else activation_light_color),
-                                       center=(light_radius, light_radius),
-                                       radius=light_radius)
-
-                    # Adding alpha gradient to the circle:
-                    light_array = pygame.surfarray.pixels_alpha(light_surface)
-
-                    y = np.linspace(-1, 1, light_array.shape[1])[None, :] * 255
-                    x = np.linspace(-1, 1, light_array.shape[0])[:, None] * 255
-
-                    alpha = np.sqrt(x ** 2 + y ** 2)
-                    alpha = 255 - np.clip(0, 255, alpha)
-                    light_array[:] = alpha[:]
-
-                    del light_array
-                    light_surface.unlock()
-
-                    # Drawing the generated surface on the main surface:
-                    surface.blit(source=light_surface,
-                                 dest=(nodes_pos[node.id][0] - light_radius,
-                                       nodes_pos[node.id][1] - light_radius))
-
-        return (surface if not return_rgb_array
-                else pygame.surfarray.array3d(surface))
-
-    def visualize(self,
-                  layout_name: str = "columns",
-                  layout_kwargs: Optional[Dict[str, Any]] = None,
-                  show: bool = True,
-                  block_thread: bool = True,
-                  save_to: Optional[str] = None,
-                  save_transparent: bool = False,
-                  figsize: Tuple[int, int] = (10, 6),
-                  node_size: int = 300,
-                  pad: int = 1,
-                  legends: bool = True,
-                  nodes_ids: bool = True,
-                  node_id_color: str = "black",
-                  edge_curviness: float = 0.1,
-                  edges_ids: bool = False,
-                  edge_id_color: str = "black",
-                  background_color: str = "snow",
-                  legend_box_color: str = "honeydew",
-                  input_color: str = "deepskyblue",
-                  output_color: str = "mediumseagreen",
-                  hidden_color: str = "silver",
-                  bias_color: str = "khaki",
-                  **kwargs,  # pylint: disable=unused-argument
-    ) -> None:
-        """ Plots the neural network (phenotype) encoded by the genome.
-
-        The network is drawn as a graph, with nodes and edges. An edge's color
-        is chosen according to the edge's weight. Edges with greater weights are
-        drawn with more intense / stronger colors. Edges connecting a node to
-        itself aren't be drawn.
-
-        This method uses `NetworkX <https://github.com/networkx/networkx>`_ to
-        handle the drawings. It positions the network's nodes according to a
-        layout, whose name you can specify in the parameter ``layout_name``. The
-        currently available layouts are:
-
-        * All the standard `NetworkX's` layouts available in this
-          `link
-          <https://networkx.org/documentation/latest/reference/drawing.html#module-networkx.drawing.layout>`_;
-        * The `graphviz` layout; it's really good, but to use it you must
-          have `Graphviz-Dev` and `pygraphviz` installed on your machine;
-        * The `columns` layout (used by default), implemented exclusively for
-          `NEvoPy`; it positions the nodes in columns (see
-          :meth:`.NeatGenome.columns_graph_layout`, specially the parameter
-          ``ideal_h_nodes_per_col``).
-
-        For the colors parameters, it's possible to pass a string with the color
-        HEX value or a string with the color's name (names available here:
-        https://matplotlib.org/3.1.0/gallery/color/named_colors.html).
-
-        Args:
-            layout_name (str): The name of the layout to be used to position the
-                network's nodes.
-            layout_kwargs (Optional[Dict[str, Any]]): Keyed arguments to be
-                passed to the layout. Check each layout documentation for more
-                information about the accepted arguments.
-            show (bool): Whether to show the generated image or not. If True, a
-                window will be created by `matplotlib` to show the image.
-            block_thread (bool): Whether to block the execution's thread while
-                showing the image. Useful for visualizing multiple networks at
-                once. In this case, you should call :meth:`.visualize` with this
-                parameter set to `False` on all genomes except for the last one,
-                so all the windows are shown simultaneously.
-            save_to (Optional[str]): Path to save the image. If `None`, the
-                image won't be automatically saved.
-            save_transparent: Whether the saved image should have a transparent
-                background or not.
-            figsize (Tuple[int, int]): Size of the matplotlib figure.
-            node_size (int): Size of the drawn nodes, in `points**2` (the area
-                of each node). Default size is 300. See the parameter ``s`` of
-                `matplotlib.axes.Axes.scatter
-                <https://matplotlib.org/3.3.3/api/_as_gen/matplotlib.axes.Axes.scatter.html>`_
-                for more information.
-            pad (int): The image's padding (distance between the figure of the
-                network and the image's border).
-            legends (bool): If `True`, a box with legends describing the nodes
-                colors will be drawn.
-            nodes_ids (bool): If `True`, the nodes will have their ID drawn
-                inside them.
-            node_id_color (str): Color of the drawn nodes ids.
-            edge_curviness (float): Angle, in radians, of the edges arcs. A
-                value of 0 indicates a straight line.
-            edges_ids (bool): If `True`, each connection/edge will have its ID
-                drawn on it. Keep in mind that some labels might overlap with
-                each other, making only one of them visible.
-            edge_id_color (str): Color of the drawn connections/edges ids.
-            background_color (str): Color of the figure's background.
-            legend_box_color (str): Color of the legend box.
-            input_color (str): Color of the input nodes.
-            output_color (str): Color of the output nodes.
-            hidden_color (str): Color of the hidden nodes.
-            bias_color (str): Color of the bias node.
-
-        Raises:
-            RuntimeError: If both `show` and `save_to` parameters are set to
-                `False` (in which case the function wouldn't be doing anything
-                but wasting computation).
-        """
-        # validating args
-        if not show and not save_to:
-            raise RuntimeError("Both \"show\" and \"save_to\" parameters are "
-                               "set to False!")
-
-        # config and start
-        plt.rcParams["axes.facecolor"] = background_color
-        graph = nx.MultiDiGraph()
-        graph.add_nodes_from([n.id for n in self.nodes()])
-        plt.figure(figsize=figsize)
-
-        if layout_kwargs is None:
-            layout_kwargs = {}
-
-        # connections
-        edges_labels = {}
-        for c in self.connections:
-            if c.enabled:
-                graph.add_edge(c.from_node.id, c.to_node.id, weight=c.weight)
-                edges_labels[(c.from_node.id, c.to_node.id)] = c.id
-
-        # selecting layout
-        if layout_name == "graphviz":
-            try:
-                # pylint: disable=import-outside-toplevel
-                # pylint: disable=unused-import
-                import pygraphviz
-                from networkx.drawing.nx_agraph import graphviz_layout
-            except ModuleNotFoundError as e:
-                raise ModuleNotFoundError(
-                    "Couldn't find the package `pygraphviz`!\nTo draw the "
-                    "genome's neural network, this package is required. To "
-                    "install it, however, you first need to install the dev "
-                    "version of `Graphviz` (https://graphviz.org/download/) on "
-                    "your system. On Ubuntu, you can do that by executing the "
-                    "following command:\n"
-                    "\t~$ sudo apt-get install -y graphviz-dev\n"
-                    "After installing `Graphviz`, just use pip to install "
-                    "`pygraphviz` and you're all set:\n"
-                    "\t~$ pip install pygraphviz") from e
-
-            if "prog" not in layout_kwargs:
-                layout_kwargs["prog"] = "dot"
-            if "args" not in layout_kwargs:
-                layout_kwargs["args"] = "-Grankdir=LR"
-            pos = graphviz_layout(graph, **layout_kwargs)
-
-        elif layout_name == "columns":
-            plt.xlim(0, figsize[0])
-            plt.ylim(0, figsize[1])
-
-            dpi = plt.gcf().dpi
-            pos = self.columns_graph_layout(*figsize,
-                                            node_size=(node_size ** 0.5) / dpi,
-                                            **layout_kwargs)
-        else:
-            nx_layout_func = getattr(nx, layout_name)
-            pos = nx_layout_func(graph, **layout_kwargs)
-
-        # plotting
-        nx.draw_networkx_nodes(graph, pos=pos,
-                               nodelist=[n.id for n in self._input_nodes],
-                               node_size=[node_size] * len(self._input_nodes),
-                               node_color=input_color, label="Input nodes")
-        nx.draw_networkx_nodes(graph, pos=pos,
-                               nodelist=[n.id for n in self._output_nodes],
-                               node_size=[node_size] * len(self._output_nodes),
-                               node_color=output_color, label="Output nodes")
-        nx.draw_networkx_nodes(graph, pos=pos,
-                               nodelist=[n.id for n in self.hidden_nodes],
-                               node_size=[node_size] * len(self.hidden_nodes),
-                               node_color=hidden_color, label="Hidden nodes")
-
-        if self._bias_node is not None:
-            nx.draw_networkx_nodes(graph, pos=pos,
-                                   nodelist=[self._bias_node.id],
-                                   node_size=[node_size],
-                                   node_color=bias_color, label="Bias node")
-
-        if graph.number_of_edges() > 0:
-            # calculating edges colors
-            edges_weights = list(nx.get_edge_attributes(graph,
-                                                        "weight").values())
-            if len(edges_weights) == 1:
-                edges_colors = [(1, 0.5, 0, 1)]
-            else:
-                min_w, max_w = np.min(edges_weights), np.max(edges_weights)
-                edges_colors = [(1, 0.6 * (1 - (w - min_w) / (max_w - min_w)),
-                                 0, 0.3 + 0.7 * (w - min_w) / (max_w - min_w))
-                                for w in edges_weights]
-
-            # drawing edges
-            nx.draw_networkx_edges(
-                graph,
-                pos=pos,
-                edge_color=edges_colors,
-                connectionstyle=f"arc3, rad={edge_curviness}"
-            )
-
-        if edges_ids:
-            nx.draw_networkx_edge_labels(graph, pos, edge_labels=edges_labels,
-                                         font_color=edge_id_color)
-
-        if nodes_ids:
-            nx.draw_networkx_labels(graph, pos,
-                                    labels={k.id: k.id
-                                            for k in self.nodes()},
-                                    font_size=10,
-                                    font_color=node_id_color,
-                                    font_family="sans-serif")
-        if legends:
-            plt.legend(facecolor=legend_box_color, borderpad=0.8,
-                       labelspacing=0.5)
-
-        plt.tight_layout(pad=pad)
-        if save_to is not None:
-            plt.savefig(save_to, transparent=save_transparent)
-
-        if show:
-            plt.show(block=block_thread)
+        return ne.neat.visualize_activations(genome=self, **kwargs)
 
 
 def _debug_mating(genes, c, gen1, gen2, new_gen):
@@ -1526,10 +870,10 @@ class FixTopNeatGenome(NeatGenome):
     """
 
     def __init__(self,
-                 fito_genome: FixedTopologyGenome,
+                 fito_genome: "ne.fixed_topology.FixedTopologyGenome",
                  num_neat_inputs: int,
                  num_neat_outputs: int,
-                 config: NeatConfig,
+                 config: "ne.neat.config.NeatConfig",
                  initial_neat_connections: bool = True) -> None:
         super().__init__(num_inputs=num_neat_inputs,
                          num_outputs=num_neat_outputs,
